backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]==0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.1
email-validator>=2.0.0

# Database & Vector Store (Supabase with pgvector)
supabase>=2.12.0
pgvector==0.2.4
sqlalchemy==2.0.25
psycopg2-binary>=2.9.11
alembic==1.13.1

# LLM & AI (Coding Interviews - Ollama only, no PyTorch!)
langchain>=0.3.0
langchain-community>=0.3.0
ollama>=0.4.7

# Document Processing (for Resume Matching & Test Evaluation)
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow>=11.0.0
pdf2image==1.16.3

# ============================================================================
# OCR & ML Dependencies - REMOVED (Not needed for Coding Interviews)
# ============================================================================
# If you need Test Evaluation (handwritten answer sheets) or Resume Matching,
# uncomment these lines and run be.bat to install PyTorch:
#
# paddleocr>=2.7.0
# paddlepaddle>=2.5.0
# sentence-transformers==2.3.1
# transformers>=4.30.0
# torch - Installed by be.bat with CUDA support if needed
# torchvision - Installed by be.bat with CUDA support if needed
# ============================================================================

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi>=0.1.9

# Utilities
python-dotenv==1.0.0
httpx>=0.27.0
aiofiles==23.2.1

# Background Tasks (optional)
redis==5.0.1
celery==5.3.4

# Excel Export (for Voice Screening)
pandas>=2.0.0
openpyxl>=3.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
